+++ R/02-modelDiagnosticsPlot.R
@@ -123,7 +123,9 @@ modelDiagnosticsPlot <- function(input, output, session, model, values) {
         plotTargets,
         plotParams()
       ) %>%
-        shinyTryCatch(errorTitle = "Error in plot", alertStyle = "shinyalert")
+        shinyTryCatch(errorTitle = "Error during plotting",
+                      warningTitle = "Warning during plotting",
+                      alertStyle = "shinyalert")
       
       # we need to trigger the update after pressing "Apply", that's why we use the if condition
       if (input$applyRangesDiag > 0) {
@@ -148,7 +150,9 @@ modelDiagnosticsPlot <- function(input, output, session, model, values) {
       # this only works with ggplots when print() is used 
       plotFunTargetDiagnostics()() %>%
         print() %>%
-        shinyTryCatch(errorTitle = "Error in plot", alertStyle = "shinyalert")
+        shinyTryCatch(errorTitle = "Error during plotting",
+                      warningTitle = "Warning during plotting",
+                      alertStyle = "shinyalert")
     },
     cacheKeyExpr = {
       plotParams()
@@ -177,7 +181,9 @@ modelDiagnosticsPlot <- function(input, output, session, model, values) {
         plotTargets,
         params
       ) %>%
-        shinyTryCatch(errorTitle = "Error in plot", alertStyle = "shinyalert")
+        shinyTryCatch(errorTitle = "Error during plotting",
+                      warningTitle = "Warning during plotting",
+                      alertStyle = "shinyalert")
     }
   })