+++ NEWS.md
@@ -1,3 +1,10 @@
+# ReSources 24.09.0
+
+## Bug Fixes
+- fix assue with the assignment of source names (#132)
+- remove warnings for deprecated functions from dplyr
+- fix old package name: MpiIsoApp -> DSSM
+
 # ReSources 24.08.2
 
 ## New Features